Refactored Globe class to the Globe.js module
